instances_macros::access::errors

Enum ValidBugs

pub(super) enum ValidBugs {

    TraitImplNotAllowed {},
    SelfTypeNotPath {},
    QSelfTypeNotGeneric {},
    NestedQSelfNotAllowed {},
    GenericParameterHasArguments {},
    QSelfGenericNotDeclared {},
    QSelfGenericNotFirst {},
    MultipleAssociatedSegments {},
    AssociatedTypeHasArguments {},
    AssociatedTypeMissing {},
    FirstTypeNotGeneric {},
    SelfNotFirstPathSegment {},
    SelfHasMultipleFollowingSegments {},
    SelfUsedInQualifiedSelfPath {},
    SelfHasGenericArguments {},
    DuplicateImplGenericParameter {},
    NodeBoundGenericArgNotGenericParam {},
    NodeBoundGenericArgNotExpectedGeneric {},
    NodeBoundIntermediatePathHasArgs {},
    ExtractedBoundLifetimeAfterGeneric {},
    ExtractedBoundGenericOrderMismatch {},
    SelfTypeNodeBoundMismatch {},
    SelfTypeNodeBoundQualifierMismatch {},
    SelfTypeNotValidated {},
    InstanceAssocMustUseSelf {},
    InstanceQAssocMustUseSelf {},
    InstanceAssocTraitQualificationRequired {},
    InstTraitNotRemovedFromImpl {},
    AssociatedItemRequiresQualifier {},
    InstanceAssocTypeSelfReferenceNotAllowed {},
    InvalidSelfFnArgument {},
    SelfNotAllowedInConstItem {},
    InstanceAssocTypePredicateNotRemoved {},
    StandaloneSelfNotReplaced {},
    ItemNotProvidedInFile {},
    ImplProvidedInFile {},
    ImplNotProvidedInFile {},
    NotDocItemNotProvided {},
    DocItemNotProvided {},
    InvalidItemCfg {},
    SelfFoundAfterTransformation {},
    InnerSelfItem {},
    SelfAssocNotAllowedInTraitBound {},

}

Variants

TraitImplNotAllowed

SelfTypeNotPath

QSelfTypeNotGeneric

NestedQSelfNotAllowed

GenericParameterHasArguments

QSelfGenericNotDeclared

QSelfGenericNotFirst

MultipleAssociatedSegments

AssociatedTypeHasArguments

AssociatedTypeMissing

FirstTypeNotGeneric

SelfNotFirstPathSegment

SelfHasMultipleFollowingSegments

SelfUsedInQualifiedSelfPath

SelfHasGenericArguments

DuplicateImplGenericParameter

NodeBoundGenericArgNotGenericParam

NodeBoundGenericArgNotExpectedGeneric

NodeBoundIntermediatePathHasArgs

ExtractedBoundLifetimeAfterGeneric

ExtractedBoundGenericOrderMismatch

SelfTypeNodeBoundMismatch

SelfTypeNodeBoundQualifierMismatch

SelfTypeNotValidated

InstanceAssocMustUseSelf

InstanceQAssocMustUseSelf

InstanceAssocTraitQualificationRequired

InstTraitNotRemovedFromImpl

AssociatedItemRequiresQualifier

InstanceAssocTypeSelfReferenceNotAllowed

InvalidSelfFnArgument

SelfNotAllowedInConstItem

InstanceAssocTypePredicateNotRemoved

StandaloneSelfNotReplaced

ItemNotProvidedInFile

ImplProvidedInFile

ImplNotProvidedInFile

NotDocItemNotProvided

DocItemNotProvided

InvalidItemCfg

SelfFoundAfterTransformation

InnerSelfItem

SelfAssocNotAllowedInTraitBound

Implementations

impl ValidBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.