instances_macros

Module args

Parsers for instances proc-macros user inputs.

This module defines the argument parsers used by the instance proc macros. Each parser converts an incoming TokenStream into a semantically meaningful representation tailored to the macro being expanded.

Structs

InstanceAccessArgs 🔒

Parsed input for the instance_access proc-macro.

InstanceDirectAccessArgs 🔒

Parsed input for the instance_direct_access proc-macro.

InstanceGetterArgs 🔒

Parsed input for the instance_get_access proc-macro.

InstanceImplArgs 🔒

Parsed input for the instance_impl proc-macro which follows instance proc-macro applied on its associated trait definition.

InstanceTraitArgs 🔒

Parsed input for the instance_trait proc-macro.

Enums

InstanceArgs 🔒

Arguments accepted by instance-related attributes on a trait definition.

InstanceNodeArgs 🔒

Parsed input for the instance_node proc-macro.