instances_macros::impls::addon

Function extract_generic_idents_from_ty

pub(crate) fn extract_generic_idents_from_ty(
    ty: &Type,
    generics: &Generics,
) -> Punctuated<GenericParam, Comma>

Extracts the generic parameters referenced by a given Self type.

Only generic identifiers appearing in Self are retained. Bounds, defaults, and generics used solely by the implemented trait are omitted.

impl<'a, T, R: Marker, const B: bool> Example<...> for Phantom<&'a R> {}

yields:

<'a, R>