instances_macros::impls::counters

Type Alias CounterArgs

pub(crate) type CounterArgs = Vec<CounterArg>;

A vector of CounterArg values describing all instance-counter const-generic arguments supplied by an implementation.

Each CounterArg represents a single const-generic argument from the implemented trait path and is identified exclusively by its zero-based CounterArg::generic_index.

Ordering semantics

The order of this vector is intentional and user-defined. It does not need to be ascending by generic_index, nor does it necessarily mirror the textual order of the trait arguments appearing in the implementation.

When instance-counter arguments are interleaved with other generic arguments, this vector preserves the order in which the user intends those counters to participate in instance resolution.

Expansion logic must therefore rely on generic_index for stable trait–impl association and treat the vector order as semantically meaningful.

Examples

trait MyTrait<const A: u8, T, const B: u8, const C: u8> {}

impl<T> MyTrait<4, T, 7, 9> for MyType {}

If the user selects counters in the order B, A, the collected arguments will be:

CounterArgs [
  CounterArg { generic_index: 2, const_lit: 7 },
  CounterArg { generic_index: 0, const_lit: 4 },
]

Likewise, when counters are interleaved with other arguments:

trait MyTrait<T, const X: u8, U, const Y: u8> {}

impl<T, U> MyTrait<T, 3, U, 8> for MyType {}

And the user specifies the instance order Y, X, the resulting vector preserves that intent:

CounterArgs [
  CounterArg { generic_index: 3, const_lit: 8 },
  CounterArg { generic_index: 1, const_lit: 3 },
]

Aliased Type

pub(crate) struct CounterArgs { /* private fields */ }