instances_macros::impls::errors

Enum AddonBugs

pub(super) enum AddonBugs {

    ExtractedIdentifierRawExprsLengthMismatch {},
    ExtractedIdentifierRawExprIsInconsistent {},
    InvalidCounterAccessImplIdent {},
    CounterAccessImplAddonNotFound {},
    InvalidGlobalTypePathForCounterAccess {},
    InvalidCounterAccessImplPathSegmentsLen {},
    InvalidCounterAccessImplPathSegment {},
    InvalidCounterAccessImplTraitSegment {},
    CounterAccessImplHasInvalidGenericArgs {},
    CounterAccessImplExpectedGenericArgsLenInvalid {},
    CounterAccessImplOriginalTypeGenericArgUnavailable {},
    CounterAccessImplOriginalTypeNotTypeGenericArg {},
    CounterAccessImplOriginalTypeGenericArgInvalid {},
    CounterAccessImplExpectedConstGenericArgsLenInvalid {},
    CounterAccessImplExpectedConstGenericFoundElse {},
    CounterAccessImplHashExprInvalid {},
    CounterAccessImplOriginalTypeGenericWhereClausesInvalid {},
    CounterAccessImplOriginalTypeGenericParamsInvalid {},
    CounterAccessImplAddonWithNoAssocItems {},
    CounterAccessImplCounterAssocNotType {},
    CounterAccessImplAddonTypeAssocIdentInvalid {},
    CounterAccessImplInvalidCounterTy {},
    CounterAccessImplHasExcessAssocs {},
    InvalidIdentHashLenAccessImplIdent {},
    IdentHashLenAccessImplAddonNotFound {},
    IdentHashLenAccessImplAddonMissing {},
    InvalidGlobalTypePathForIdentHashLenAccess {},
    InvalidIdentHashLenAccessImplPathSegmentsLen {},
    InvalidIdentHashLenAccessImplPathSegment {},
    InvalidIdentHashLenAccessImplTraitSegment {},
    IdentHashLenAccessImplHasInvalidGenericArgs {},
    IdentHashLenAccessImplExpectedGenericArgsLenInvalid {},
    IdentHashLenAccessImplOriginalTypeGenericArgUnavailable {},
    IdentHashLenAccessImplOriginalTypeNotTypeGenericArg {},
    IdentHashLenAccessImplOriginalTypeGenericArgInvalid {},
    IdentHashLenAccessImplExpectedConstGenericArgsLenInvalid {},
    IdentHashLenAccessImplExpectedConstGenericFoundElse {},
    IdentHashLenAccessImplHashExprInvalid {},
    IdentHashLenAccessImplOriginalTypeGenericWhereClausesInvalid {},
    IdentHashLenAccessImplOriginalTypeGenericParamsInvalid {},
    IdentHashLenAccessImplAddonWithNoAssocItems {},
    IdentHashLenAccessImplLengthAssocNotType {},
    IdentHashLenAccessImplAddonTypeAssocIdentInvalid {},
    IdentHashLenAccessGivenCounterIndexInvalid {},
    IdentHashLenAccessImplInvalidLengthTy {},
    IdentHashLenAccessImplHasExcessAssocs {},
    IdentHashLenAccessCounterIndexGenericArgUnavailable {},
    IdentHashLenAccessCounterIndexNotConstGenericArg {},
    IdentHashLenAccessCounterIndexNotConstExprLit {},
    IdentHashLenAccessCounterIndexNotConstLitInt {},
    InvalidTerminalAccessImplIdent {},
    InvalidGlobalTypePathForTerminalAccess {},
    InvalidTerminalAccessImplPathSegmentsLen {},
    InvalidTerminalAccessImplPathSegment {},
    InvalidTerminalAccessImplTraitSegment {},
    TerminalAccessImplHasInvalidGenericArgs {},
    TerminalAccessImplExpectedGenericArgsLenInvalid {},
    TerminalAccessImplOriginalTypeGenericArgUnavailable {},
    TerminalAccessImplOriginalTypeNotTypeGenericArg {},
    TerminalAccessImplOriginalTypeGenericArgInvalid {},
    TerminalAccessImplOriginalTypeGenericWhereClausesInvalid {},
    TerminalAccessImplOriginalTypeGenericParamsInvalid {},
    TerminalAccessImplAddonWithNoAssocItems {},
    TerminalAccessImplCounterAssocNotType {},
    TerminalAccessImplAddonTypeAssocIdentInvalid {},
    TerminalAccessImplInvalidCounterTy {},
    TerminalAccessImplHasExcessAssocs {},
    TerminalAccessImplNotFound {},
    TerminalAccessImplAttemptedForNonTerminalCounter {},
    TerminalCheckerConstAttemptedForNonTerminalCounter {},
    TerminalCheckerConstInvalidIdent {},
    TerminalCheckerConstAttemptedForGenericImpl {},
    TerminalCheckerConstInvalidExpr {},
    TerminalCheckerConstNotFound {},
    LastInstanceCounterInvalid {},
    BoundaryAccessHashOfIndexFromHashExprsUnavailable {},
    InvalidGlobalTypePathForBoundaryAccess {},
    InvalidBoundaryAccessImplPathSegmentsLen {},
    InvalidBoundaryAccessImplPathSegment {},
    InvalidBoundaryAccessImplTraitSegment {},
    BoundaryAccessImplHasInvalidGenericArgs {},
    BoundaryAccessImplOriginalTypeGenericArgUnavailable {},
    BoundaryAccessImplOriginalTypeNotTypeGenericArg {},
    BoundaryAccessImplOriginalTypeGenericArgInvalid {},
    BoundaryAccessHashExprArgsLenInvalid {},
    BoundaryAccessHashExprArgNotConst {},
    BoundaryAccessHashExprArgConstInvalid {},
    BoundaryAccessImplOriginalTypeGenericWhereClausesInvalid {},
    BoundaryAccessImplOriginalTypeGenericParamsInvalid {},
    BoundaryAccessImplAddonWithNoAssocItems {},
    BoundaryAccessImplHasExcessAssocs {},
    BoundaryAccessImplCounterAssocTypeNotFound {},
    BoundaryAccessImplAddonTypeAssocIdentInvalid {},
    BoundaryAccessImplInvalidCounterTy {},
    BoundaryAccessImplHashAssocConstNotFound {},
    BoundaryAccessImplAddonConstAssocIdentInvalid {},
    BoundaryAccessImplInvalidHashConstTy {},
    BoundaryAccessImplInvalidConstAssocExpr {},
    BoundaryAccessImplNotFound {},
    OnSetAccessHashOfIndexFromHashExprsUnavailable {},
    InvalidGlobalTypePathForOnSetAccess {},
    InvalidOnSetAccessImplPathSegmentsLen {},
    InvalidOnSetAccessImplPathSegment {},
    InvalidOnSetAccessImplTraitSegment {},
    OnSetAccessImplHasInvalidGenericArgs {},
    OnSetAccessImplOriginalTypeGenericArgUnavailable {},
    OnSetAccessImplOriginalTypeNotTypeGenericArg {},
    OnSetAccessImplOriginalTypeGenericArgInvalid {},
    OnSetAccessHashExprArgsLenInvalid {},
    OnSetAccessHashExprArgNotConst {},
    OnSetAccessHashExprArgConstInvalid {},
    OnSetAccessImplOriginalTypeGenericWhereClausesInvalid {},
    OnSetAccessImplOriginalTypeGenericParamsInvalid {},
    OnSetAccessImplAddonWithNoAssocItems {},
    OnSetAccessImplHasExcessAssocs {},
    OnSetAccessImplCounterAssocTypeNotFound {},
    OnSetAccessImplAddonTypeAssocIdentInvalid {},
    OnSetAccessImplInvalidCounterTy {},
    OnSetAccessImplHashAssocConstNotFound {},
    OnSetAccessImplAddonConstAssocIdentInvalid {},
    OnSetAccessImplInvalidHashConstTy {},
    OnSetAccessImplInvalidConstAssocExpr {},
    OnSetAccessImplNotFound {},

}

Variants

ExtractedIdentifierRawExprsLengthMismatch

ExtractedIdentifierRawExprIsInconsistent

InvalidCounterAccessImplIdent

CounterAccessImplAddonNotFound

InvalidGlobalTypePathForCounterAccess

InvalidCounterAccessImplPathSegmentsLen

InvalidCounterAccessImplPathSegment

InvalidCounterAccessImplTraitSegment

CounterAccessImplHasInvalidGenericArgs

CounterAccessImplExpectedGenericArgsLenInvalid

CounterAccessImplOriginalTypeGenericArgUnavailable

CounterAccessImplOriginalTypeNotTypeGenericArg

CounterAccessImplOriginalTypeGenericArgInvalid

CounterAccessImplExpectedConstGenericArgsLenInvalid

CounterAccessImplExpectedConstGenericFoundElse

CounterAccessImplHashExprInvalid

CounterAccessImplOriginalTypeGenericWhereClausesInvalid

CounterAccessImplOriginalTypeGenericParamsInvalid

CounterAccessImplAddonWithNoAssocItems

CounterAccessImplCounterAssocNotType

CounterAccessImplAddonTypeAssocIdentInvalid

CounterAccessImplInvalidCounterTy

CounterAccessImplHasExcessAssocs

InvalidIdentHashLenAccessImplIdent

IdentHashLenAccessImplAddonNotFound

IdentHashLenAccessImplAddonMissing

InvalidGlobalTypePathForIdentHashLenAccess

InvalidIdentHashLenAccessImplPathSegmentsLen

InvalidIdentHashLenAccessImplPathSegment

InvalidIdentHashLenAccessImplTraitSegment

IdentHashLenAccessImplHasInvalidGenericArgs

IdentHashLenAccessImplExpectedGenericArgsLenInvalid

IdentHashLenAccessImplOriginalTypeGenericArgUnavailable

IdentHashLenAccessImplOriginalTypeNotTypeGenericArg

IdentHashLenAccessImplOriginalTypeGenericArgInvalid

IdentHashLenAccessImplExpectedConstGenericArgsLenInvalid

IdentHashLenAccessImplExpectedConstGenericFoundElse

IdentHashLenAccessImplHashExprInvalid

IdentHashLenAccessImplOriginalTypeGenericWhereClausesInvalid

IdentHashLenAccessImplOriginalTypeGenericParamsInvalid

IdentHashLenAccessImplAddonWithNoAssocItems

IdentHashLenAccessImplLengthAssocNotType

IdentHashLenAccessImplAddonTypeAssocIdentInvalid

IdentHashLenAccessGivenCounterIndexInvalid

IdentHashLenAccessImplInvalidLengthTy

IdentHashLenAccessImplHasExcessAssocs

IdentHashLenAccessCounterIndexGenericArgUnavailable

IdentHashLenAccessCounterIndexNotConstGenericArg

IdentHashLenAccessCounterIndexNotConstExprLit

IdentHashLenAccessCounterIndexNotConstLitInt

InvalidTerminalAccessImplIdent

InvalidGlobalTypePathForTerminalAccess

InvalidTerminalAccessImplPathSegmentsLen

InvalidTerminalAccessImplPathSegment

InvalidTerminalAccessImplTraitSegment

TerminalAccessImplHasInvalidGenericArgs

TerminalAccessImplExpectedGenericArgsLenInvalid

TerminalAccessImplOriginalTypeGenericArgUnavailable

TerminalAccessImplOriginalTypeNotTypeGenericArg

TerminalAccessImplOriginalTypeGenericArgInvalid

TerminalAccessImplOriginalTypeGenericWhereClausesInvalid

TerminalAccessImplOriginalTypeGenericParamsInvalid

TerminalAccessImplAddonWithNoAssocItems

TerminalAccessImplCounterAssocNotType

TerminalAccessImplAddonTypeAssocIdentInvalid

TerminalAccessImplInvalidCounterTy

TerminalAccessImplHasExcessAssocs

TerminalAccessImplNotFound

TerminalAccessImplAttemptedForNonTerminalCounter

TerminalCheckerConstAttemptedForNonTerminalCounter

TerminalCheckerConstInvalidIdent

TerminalCheckerConstAttemptedForGenericImpl

TerminalCheckerConstInvalidExpr

TerminalCheckerConstNotFound

LastInstanceCounterInvalid

BoundaryAccessHashOfIndexFromHashExprsUnavailable

InvalidGlobalTypePathForBoundaryAccess

InvalidBoundaryAccessImplPathSegmentsLen

InvalidBoundaryAccessImplPathSegment

InvalidBoundaryAccessImplTraitSegment

BoundaryAccessImplHasInvalidGenericArgs

BoundaryAccessImplOriginalTypeGenericArgUnavailable

BoundaryAccessImplOriginalTypeNotTypeGenericArg

BoundaryAccessImplOriginalTypeGenericArgInvalid

BoundaryAccessHashExprArgsLenInvalid

BoundaryAccessHashExprArgNotConst

BoundaryAccessHashExprArgConstInvalid

BoundaryAccessImplOriginalTypeGenericWhereClausesInvalid

BoundaryAccessImplOriginalTypeGenericParamsInvalid

BoundaryAccessImplAddonWithNoAssocItems

BoundaryAccessImplHasExcessAssocs

BoundaryAccessImplCounterAssocTypeNotFound

BoundaryAccessImplAddonTypeAssocIdentInvalid

BoundaryAccessImplInvalidCounterTy

BoundaryAccessImplHashAssocConstNotFound

BoundaryAccessImplAddonConstAssocIdentInvalid

BoundaryAccessImplInvalidHashConstTy

BoundaryAccessImplInvalidConstAssocExpr

BoundaryAccessImplNotFound

OnSetAccessHashOfIndexFromHashExprsUnavailable

InvalidGlobalTypePathForOnSetAccess

InvalidOnSetAccessImplPathSegmentsLen

InvalidOnSetAccessImplPathSegment

InvalidOnSetAccessImplTraitSegment

OnSetAccessImplHasInvalidGenericArgs

OnSetAccessImplOriginalTypeGenericArgUnavailable

OnSetAccessImplOriginalTypeNotTypeGenericArg

OnSetAccessImplOriginalTypeGenericArgInvalid

OnSetAccessHashExprArgsLenInvalid

OnSetAccessHashExprArgNotConst

OnSetAccessHashExprArgConstInvalid

OnSetAccessImplOriginalTypeGenericWhereClausesInvalid

OnSetAccessImplOriginalTypeGenericParamsInvalid

OnSetAccessImplAddonWithNoAssocItems

OnSetAccessImplHasExcessAssocs

OnSetAccessImplCounterAssocTypeNotFound

OnSetAccessImplAddonTypeAssocIdentInvalid

OnSetAccessImplInvalidCounterTy

OnSetAccessImplHashAssocConstNotFound

OnSetAccessImplAddonConstAssocIdentInvalid

OnSetAccessImplInvalidHashConstTy

OnSetAccessImplInvalidConstAssocExpr

OnSetAccessImplNotFound

Implementations

impl AddonBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.