instances_macros::node::errors

Enum AccessBugs

pub(super) enum AccessBugs {

    CountersGenericsIndexesEmpty {},
    TraitPathNotAngleArgs {},
    AccessTyGenericArgNotFound {},
    AccessGenericArgNotType {},
    UnexpectedAccessTyGenericArg {},
    UnexpectedAccessTraitBound {},
    ExactAccessGenBoundQualifierNotFound {},
    ExactAccessGenBoundQualifierNotSupportCrate {},
    ExactAccessGenBoundIdentNotFound {},
    ExactAccessGenBoundIdentInvalid {},
    ExactAccessGenBoundArgsNotAngle {},
    ExactAccessGenBoundArgsInconsistent {},
    ExactAccessGenBoundArgNotType {},
    ExactAccessGenBoundArgNotTypePath {},
    ExactAccessGenBoundArgTypeQSelfNotFound {},
    ExactAccessGenBoundArgTypeQSelfInvalid {},
    ExactAccessGenBoundArgTypePathProjectionNotFound {},
    ExactAccessGenBoundArgTypePathProjectionInvalid {},
    ExactAccessBoundQualifierNotFound {},
    ExactAccessBoundQualifierInvalid {},
    ExactAccessGenBoundPathNotFound {},
    ExactAccessBoundPathNotFound {},
    ExactAccessGenBoundPathInconsistent {},

}

Variants

CountersGenericsIndexesEmpty

TraitPathNotAngleArgs

AccessTyGenericArgNotFound

AccessGenericArgNotType

UnexpectedAccessTyGenericArg

UnexpectedAccessTraitBound

ExactAccessGenBoundQualifierNotFound

ExactAccessGenBoundQualifierNotSupportCrate

ExactAccessGenBoundIdentNotFound

ExactAccessGenBoundIdentInvalid

ExactAccessGenBoundArgsNotAngle

ExactAccessGenBoundArgsInconsistent

ExactAccessGenBoundArgNotType

ExactAccessGenBoundArgNotTypePath

ExactAccessGenBoundArgTypeQSelfNotFound

ExactAccessGenBoundArgTypeQSelfInvalid

ExactAccessGenBoundArgTypePathProjectionNotFound

ExactAccessGenBoundArgTypePathProjectionInvalid

ExactAccessBoundQualifierNotFound

ExactAccessBoundQualifierInvalid

ExactAccessGenBoundPathNotFound

ExactAccessBoundPathNotFound

ExactAccessGenBoundPathInconsistent

Implementations

impl AccessBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.