instances_macros::node::errors

Enum StateBugs

pub(super) enum StateBugs {

    FetchedImplItemNotTypeAssoc {},
    TraitItemIdxFetchFailed {},
    ImplItemIdxFetchFailed {},
    FetchedTraitItemNotTypeAssoc {},
    FirstCounterNotFound {},
    FoundNoBoundForInstance {},
    RequiresBoundAsInstanceTraitBound {},
    ExpectedInstanceTraitBoundInvalid {},
    HrtbInstanceTraitsUnavailable {},
    InstanceTraitBoundLastSegUnavailable {},
    InstanceTraitBoundNotAngleArgs {},
    InstanceTraitBoundInstanceTypeNumArgUnavailable {},
    InstanceTraitBoundInstanceTypeNumArgNotType {},
    GlobalNodeInvalidGenerics {},
    GlobalNodeNotFound {},
    TwinNodeInvalidGenerics {},
    TwinNodeNotFound {},
    CheckerNodeInvalidGenerics {},
    CheckerNodeNotFound {},
    InitialNodeInvalidGenerics {},
    InitialNodeNotFound {},
    FinalNodeInvalidGenerics {},
    FinalNodeNotFound {},
    TerminalGlobalNodeInvalidGenerics {},
    TerminalGlobalNodeInvalidType {},
    TerminalGlobalNodeNotFound {},
    TerminalTwinNodeInvalidGenerics {},
    TerminalTwinNodeInvalidType {},
    TerminalTwinNodeNotFound {},
    TerminalCheckerNodeInvalidGenerics {},
    TerminalCheckerNodeNotFound {},
    ValidateContextFirstOfIndexesUnavailable {},
    StateInstanceNodeUnavailable {},
    StateInstanceNodeBoundUnavailable {},
    StateInstanceNodeReplacementBoundIdentInvalid {},
    StateInstanceNodeReplacementBoundLastSegUnavailable {},
    StateInstanceNodeReplacementBoundNotAngleArgs {},
    StateInstanceOldNodeBoundNotTraitBound {},
    TerminalTwinBoundsMoreThanSingleTerminalBound {},
    TerminalTwinBoundNotTraitBound {},
    NonReplacedInstanceBoundLastPathSegNotFound {},
    TerminalGlobalBoundsMoreThanSingleTerminalBound {},
    TerminalGlobalBoundNotTraitBound {},
    TerminalCheckerDefaultExprNotFound {},
    TerminalCheckerDefaultNotExprPath {},
    TerminalCheckerDefaultExprPathQSelfNotFound {},
    TerminalCheckerDefaultExprPathQSelfInvalid {},
    TerminalCheckerDefaultExprPathLastSegNotFound {},
    TerminalCheckerDefaultExprPathLastSegInvalid {},
    TerminalCheckerDefaultExprPathQSelfTraitBoundInvalid {},
    ImplInstanceTyLastSegNotFound {},
    ImplTerminalTwinTyNotTypePath {},
    ImplTerminalTwinTyQSelfNotFound {},
    ImplTerminalTwinTyQSelfInvalid {},
    ImplTerminalTwinTyLastSegNotFound {},
    ImplTerminalTwinTyLastSegInvalid {},
    ImplTerminalTwinTyLastQSelfBoundInvalid {},
    ImplTerminalGlobalTyNotTypePath {},
    ImplTerminalGlobalTyQSelfNotFound {},
    ImplTerminalGlobalTyQSelfInvalid {},
    ImplTerminalGlobalTyLastSegNotFound {},
    ImplTerminalGlobalTyLastSegInvalid {},
    ImplTerminalGlobalTyLastQSelfBoundInvalid {},

}

Variants

FetchedImplItemNotTypeAssoc

TraitItemIdxFetchFailed

ImplItemIdxFetchFailed

FetchedTraitItemNotTypeAssoc

FirstCounterNotFound

FoundNoBoundForInstance

RequiresBoundAsInstanceTraitBound

ExpectedInstanceTraitBoundInvalid

HrtbInstanceTraitsUnavailable

InstanceTraitBoundLastSegUnavailable

InstanceTraitBoundNotAngleArgs

InstanceTraitBoundInstanceTypeNumArgUnavailable

InstanceTraitBoundInstanceTypeNumArgNotType

GlobalNodeInvalidGenerics

GlobalNodeNotFound

TwinNodeInvalidGenerics

TwinNodeNotFound

CheckerNodeInvalidGenerics

CheckerNodeNotFound

InitialNodeInvalidGenerics

InitialNodeNotFound

FinalNodeInvalidGenerics

FinalNodeNotFound

TerminalGlobalNodeInvalidGenerics

TerminalGlobalNodeInvalidType

TerminalGlobalNodeNotFound

TerminalTwinNodeInvalidGenerics

TerminalTwinNodeInvalidType

TerminalTwinNodeNotFound

TerminalCheckerNodeInvalidGenerics

TerminalCheckerNodeNotFound

ValidateContextFirstOfIndexesUnavailable

StateInstanceNodeUnavailable

StateInstanceNodeBoundUnavailable

StateInstanceNodeReplacementBoundIdentInvalid

StateInstanceNodeReplacementBoundLastSegUnavailable

StateInstanceNodeReplacementBoundNotAngleArgs

StateInstanceOldNodeBoundNotTraitBound

TerminalTwinBoundsMoreThanSingleTerminalBound

TerminalTwinBoundNotTraitBound

NonReplacedInstanceBoundLastPathSegNotFound

TerminalGlobalBoundsMoreThanSingleTerminalBound

TerminalGlobalBoundNotTraitBound

TerminalCheckerDefaultExprNotFound

TerminalCheckerDefaultNotExprPath

TerminalCheckerDefaultExprPathQSelfNotFound

TerminalCheckerDefaultExprPathQSelfInvalid

TerminalCheckerDefaultExprPathLastSegNotFound

TerminalCheckerDefaultExprPathLastSegInvalid

TerminalCheckerDefaultExprPathQSelfTraitBoundInvalid

ImplInstanceTyLastSegNotFound

ImplTerminalTwinTyNotTypePath

ImplTerminalTwinTyQSelfNotFound

ImplTerminalTwinTyQSelfInvalid

ImplTerminalTwinTyLastSegNotFound

ImplTerminalTwinTyLastSegInvalid

ImplTerminalTwinTyLastQSelfBoundInvalid

ImplTerminalGlobalTyNotTypePath

ImplTerminalGlobalTyQSelfNotFound

ImplTerminalGlobalTyQSelfInvalid

ImplTerminalGlobalTyLastSegNotFound

ImplTerminalGlobalTyLastSegInvalid

ImplTerminalGlobalTyLastQSelfBoundInvalid

Implementations

impl StateBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.