instances_macros::traits::counters

Type Alias CounterParams

pub(crate) type CounterParams = Vec<CounterParam>;

A vector of CounterParam values describing all instance-counter const-generic parameters declared on the instance trait.

Each CounterParam represents a single const-generic parameter from the trait’s generic parameter list and is identified exclusively by its zero-based CounterParam::generic_index.

Ordering semantics

The order of this vector is intentional and user-defined.
It does not need to be ascending by generic_index, nor does it necessarily mirror the textual order of the trait’s generic parameters.

When instance-counter const generics are mixed with other generic parameters in the trait definition, this vector preserves the order in which the user intends those counters to participate in instance resolution.

Expansion logic must therefore rely on generic_index for stable trait–impl association and treat the vector order as semantically meaningful.

Examples

// Trait definition
trait MyTrait<const A: u8, T, const B: u8, const C: u8> {}

If the user declares instance counters in the order B, A, the collected parameters will be:

CounterParams [
  CounterParam { generic_index: 2, const_param: B },
  CounterParam { generic_index: 0, const_param: A },
]

Likewise, if the user interleaves counters with other generics:

trait MyTrait<T, const X: u8, U, const Y: u8> {}

And specifies the instance order Y, X, the resulting vector preserves that intent:

CounterParams [
  CounterParam { generic_index: 3, const_param: Y },
  CounterParam { generic_index: 1, const_param: X },
]

Aliased Type

pub(crate) struct CounterParams { /* private fields */ }