instances_macros::traits::errors

Enum AffiliateBugs

pub(super) enum AffiliateBugs {

    AffiliateCounterNotFound {},
    ReverseAffiliateCountersMissing {},
    ReverseAffiliateCountersHasGenerics {},
    ReverseAffiliateCountersInvalidIdent {},
    ReverseAffiliateCountersBoundNotFound {},
    ReverseAffiliateInstanceMissing {},
    ReverseAffiliateInstanceHasGenerics {},
    ReverseAffiliateInstanceInvalidIdent {},
    ReverseAffiliateInstanceBoundNotFound {},
    ReverseAffiliateInstanceBoundInvalid {},
    CountersForReverseAffiliatePredicateMissing {},
    CountersForReverseAffiliatePredicateHasLifetimes {},
    CountersForReverseAffiliateNotTypePredicate {},
    CountersForReverseAffiliatePredicateWrongBoundedTy {},
    CountersForReverseAffiliatePredicateBoundNotFound {},
    PostReverseAffiliateTraitWhereClauseMissing {},
    NextAffiliateCountersMissing {},
    NextAffiliateCountersHasGenerics {},
    NextAffiliateCountersInvalidIdent {},
    NextAffiliateCountersBoundNotFound {},
    NextAffiliateInstanceMissing {},
    NextAffiliateInstanceHasGenerics {},
    NextAffiliateInstanceInvalidIdent {},
    NextAffiliateInstanceBoundNotFound {},
    NextAffiliateInstanceBoundInvalid {},
    PostNextAffiliateTraitWhereClauseMissing {},
    BackAffiliateCountersMissing {},
    BackAffiliateCountersHasGenerics {},
    BackAffiliateCountersInvalidIdent {},
    BackAffiliateCountersBoundNotFound {},
    BackAffiliateInstanceMissing {},
    BackAffiliateInstanceHasGenerics {},
    BackAffiliateInstanceInvalidIdent {},
    BackAffiliateInstanceBoundNotFound {},
    BackAffiliateInstanceBoundInvalid {},
    PostBackAffiliateTraitWhereClauseMissing {},
    FloorAffiliatesCountersMissing {},
    FloorAffiliatesCountersHasGenerics {},
    FloorAffiliatesCountersInvalidIdent {},
    FloorAffiliatesCountersBoundNotFound {},
    FloorAffiliatesInstancesMissing {},
    FloorAffiliatesInstancesHasGenerics {},
    FloorAffiliatesInstancesInvalidIdent {},
    FloorAffiliatesInstancesBoundNotFound {},
    FloorAffiliatesInstancesBoundInvalid {},
    PostFloorAffiliatesTraitWhereClauseMissing {},
    CeilAffiliateCountersMissing {},
    CeilAffiliateCountersHasGenerics {},
    CeilAffiliateCountersInvalidIdent {},
    CeilAffiliateCountersBoundNotFound {},
    CeilAffiliateInstanceMissing {},
    CeilAffiliateInstanceHasGenerics {},
    CeilAffiliateInstanceInvalidIdent {},
    CeilAffiliateInstanceBoundNotFound {},
    CeilAffiliateInstanceBoundInvalid {},
    CountersForCeilAffiliatePredicateMissing {},
    CountersForCeilAffiliatePredicateHasLifetimes {},
    CountersForCeilAffiliateNotTypePredicate {},
    CountersForCeilAffiliatePredicateWrongBoundedTy {},
    CountersForCeilAffiliatePredicateBoundNotFound {},
    PostCeilAffiliateTraitWhereClauseMissing {},
    AffiliateCountersCheckerNotFound {},
    AffiliateCountersCheckerWrongIdent {},
    AffiliateCountersCheckerNotUnitType {},
    AffiliateCountersCheckerHasGenerics {},
    AffiliateCountersCheckerHasExpr {},

}

Variants

AffiliateCounterNotFound

ReverseAffiliateCountersMissing

ReverseAffiliateCountersHasGenerics

ReverseAffiliateCountersInvalidIdent

ReverseAffiliateCountersBoundNotFound

ReverseAffiliateInstanceMissing

ReverseAffiliateInstanceHasGenerics

ReverseAffiliateInstanceInvalidIdent

ReverseAffiliateInstanceBoundNotFound

ReverseAffiliateInstanceBoundInvalid

CountersForReverseAffiliatePredicateMissing

CountersForReverseAffiliatePredicateHasLifetimes

CountersForReverseAffiliateNotTypePredicate

CountersForReverseAffiliatePredicateWrongBoundedTy

CountersForReverseAffiliatePredicateBoundNotFound

PostReverseAffiliateTraitWhereClauseMissing

NextAffiliateCountersMissing

NextAffiliateCountersHasGenerics

NextAffiliateCountersInvalidIdent

NextAffiliateCountersBoundNotFound

NextAffiliateInstanceMissing

NextAffiliateInstanceHasGenerics

NextAffiliateInstanceInvalidIdent

NextAffiliateInstanceBoundNotFound

NextAffiliateInstanceBoundInvalid

PostNextAffiliateTraitWhereClauseMissing

BackAffiliateCountersMissing

BackAffiliateCountersHasGenerics

BackAffiliateCountersInvalidIdent

BackAffiliateCountersBoundNotFound

BackAffiliateInstanceMissing

BackAffiliateInstanceHasGenerics

BackAffiliateInstanceInvalidIdent

BackAffiliateInstanceBoundNotFound

BackAffiliateInstanceBoundInvalid

PostBackAffiliateTraitWhereClauseMissing

FloorAffiliatesCountersMissing

FloorAffiliatesCountersHasGenerics

FloorAffiliatesCountersInvalidIdent

FloorAffiliatesCountersBoundNotFound

FloorAffiliatesInstancesMissing

FloorAffiliatesInstancesHasGenerics

FloorAffiliatesInstancesInvalidIdent

FloorAffiliatesInstancesBoundNotFound

FloorAffiliatesInstancesBoundInvalid

PostFloorAffiliatesTraitWhereClauseMissing

CeilAffiliateCountersMissing

CeilAffiliateCountersHasGenerics

CeilAffiliateCountersInvalidIdent

CeilAffiliateCountersBoundNotFound

CeilAffiliateInstanceMissing

CeilAffiliateInstanceHasGenerics

CeilAffiliateInstanceInvalidIdent

CeilAffiliateInstanceBoundNotFound

CeilAffiliateInstanceBoundInvalid

CountersForCeilAffiliatePredicateMissing

CountersForCeilAffiliatePredicateHasLifetimes

CountersForCeilAffiliateNotTypePredicate

CountersForCeilAffiliatePredicateWrongBoundedTy

CountersForCeilAffiliatePredicateBoundNotFound

PostCeilAffiliateTraitWhereClauseMissing

AffiliateCountersCheckerNotFound

AffiliateCountersCheckerWrongIdent

AffiliateCountersCheckerNotUnitType

AffiliateCountersCheckerHasGenerics

AffiliateCountersCheckerHasExpr

Implementations

impl AffiliateBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.