instances_macros::traits::errors

Enum IdentTraitItemsBugs

pub(super) enum IdentTraitItemsBugs {

    CounterIdentNotFound {},
    CounterIdentWrongIdent {},
    CounterIdentWrongType {},
    CounterIdentHasGenerics {},
    CounterIdentHasExpr {},
    IndexedCounterIdentNotFound {},
    IndexedCounterIdentWrongIdent {},
    IndexedCounterIdentExprNotFound {},
    IndexedCounterIdentInvalidExpr {},
    IndexedCounterIdentWrongType {},
    IndexedCounterIdentHasGenerics {},
    CounterIdentHashNotFound {},
    CounterIdentHashWrongIdent {},
    CounterIdentHashExprNotFound {},
    CounterIdentHashInvalidExpr {},
    CounterIdentHashWrongType {},
    CounterIdentHashHasGenerics {},
    CounterIdentCollectionLenTypeNumNotFound {},
    CounterIdentCollectionLenTypeNumHasGenerics {},
    CounterIdentCollectionLenTypeNumWrongIdent {},
    CounterIdentCollectionLenTypeNumBoundNotFound {},
    CounterTypeNumCollectionLenBoundsPredicateMissing {},
    CounterTypeNumCollectionLenBoundsPredicateHasLifetimes {},
    CounterTypeNumCollectionLenBoundsNotTypePredicate {},
    CounterTypeNumCollectionLenBoundsPredicateWrongBoundedTy {},
    CounterTypeNumCollectionLenBoundsPredicateBoundNotFound {},
    InstanceTraitWhereClauseMissing {},
    AllAffiliateCountersCollectionLenBoundsNotFound {},
    AllAffiliateCountersCollectionLenBoundsHasGenerics {},
    AllAffiliateCountersCollectionLenBoundsWrongIdent {},
    AllAffiliateCountersCollectionLenBoundsBoundNotFound {},
    CounterIdentCollectionSliceNotFound {},
    CounterIdentCollectionSliceWrongIdent {},
    CounterIdentCollectionSliceWrongType {},
    CounterIdentCollectionSliceHasGenerics {},
    CounterIdentCollectionSliceHasExpr {},
    CounterIdentCollectionGenArrayNotFound {},
    CounterIdentCollectionGenArrayWrongIdent {},
    CounterIdentCollectionGenArrayWrongType {},
    CounterIdentCollectionGenArrayHasGenerics {},
    CounterIdentCollectionGenArrayHasExpr {},
    CounterIdentHashCollectionSliceNotFound {},
    CounterIdentHashCollectionSliceWrongIdent {},
    CounterIdentHashCollectionSliceWrongType {},
    CounterIdentHashCollectionSliceHasGenerics {},
    CounterIdentHashCollectionSliceHasExpr {},
    CounterIdentHashCollectionGenArrayNotFound {},
    CounterIdentHashCollectionGenArrayWrongIdent {},
    CounterIdentHashCollectionGenArrayWrongType {},
    CounterIdentHashCollectionGenArrayHasGenerics {},
    CounterIdentHashCollectionGenArrayHasExpr {},
    CounterIdentHistoricalCollectionLenConstNotFound {},
    CounterIdentHistoricalCollectionLenConstWrongIdent {},
    CounterIdentHistoricalCollectionLenConstWrongType {},
    CounterIdentHistoricalCollectionLenConstHasGenerics {},
    CounterIdentHistoricalCollectionLenConstHasExpr {},
    CounterIdentHistoricalCollectionLenTypeNumNotFound {},
    CounterIdentHistoricalCollectionLenTypeNumHasGenerics {},
    CounterIdentHistoricalCollectionLenTypeNumWrongIdent {},
    CounterIdentHistoricalCollectionLenTypeNumBoundNotFound {},
    CounterIdentHistoricalCollectionSliceNotFound {},
    CounterIdentHistoricalCollectionSliceWrongIdent {},
    CounterIdentHistoricalCollectionSliceWrongType {},
    CounterIdentHistoricalCollectionSliceHasGenerics {},
    CounterIdentHistoricalCollectionSliceHasExpr {},
    CounterIdentHistoricalCollectionGenArrayNotFound {},
    CounterIdentHistoricalCollectionGenArrayWrongIdent {},
    CounterIdentHistoricalCollectionGenArrayWrongType {},
    CounterIdentHistoricalCollectionGenArrayHasGenerics {},
    CounterIdentHistoricalCollectionGenArrayHasExpr {},
    CounterIdentHashHistoricalCollectionSliceNotFound {},
    CounterIdentHashHistoricalCollectionSliceWrongIdent {},
    CounterIdentHashHistoricalCollectionSliceWrongType {},
    CounterIdentHashHistoricalCollectionSliceHasGenerics {},
    CounterIdentHashHistoricalCollectionSliceHasExpr {},
    CounterIdentHashHistoricalCollectionGenArrayNotFound {},
    CounterIdentHashHistoricalCollectionGenArrayWrongIdent {},
    CounterIdentHashHistoricalCollectionGenArrayWrongType {},
    CounterIdentHashHistoricalCollectionGenArrayHasGenerics {},
    CounterIdentHashHistoricalCollectionGenArrayHasExpr {},
    CounterIdentExpectedHashCheckerNotFound {},
    CounterIdentExpectedHashCheckerWrongIdent {},
    CounterIdentExpectedHashCheckerNotUnitType {},
    CounterIdentExpectedHashCheckerHasGenerics {},
    CounterIdentExpectedHashCheckerHasExpr {},

}

Variants

CounterIdentNotFound

CounterIdentWrongIdent

CounterIdentWrongType

CounterIdentHasGenerics

CounterIdentHasExpr

IndexedCounterIdentNotFound

IndexedCounterIdentWrongIdent

IndexedCounterIdentExprNotFound

IndexedCounterIdentInvalidExpr

IndexedCounterIdentWrongType

IndexedCounterIdentHasGenerics

CounterIdentHashNotFound

CounterIdentHashWrongIdent

CounterIdentHashExprNotFound

CounterIdentHashInvalidExpr

CounterIdentHashWrongType

CounterIdentHashHasGenerics

CounterIdentCollectionLenTypeNumNotFound

CounterIdentCollectionLenTypeNumHasGenerics

CounterIdentCollectionLenTypeNumWrongIdent

CounterIdentCollectionLenTypeNumBoundNotFound

CounterTypeNumCollectionLenBoundsPredicateMissing

CounterTypeNumCollectionLenBoundsPredicateHasLifetimes

CounterTypeNumCollectionLenBoundsNotTypePredicate

CounterTypeNumCollectionLenBoundsPredicateWrongBoundedTy

CounterTypeNumCollectionLenBoundsPredicateBoundNotFound

InstanceTraitWhereClauseMissing

AllAffiliateCountersCollectionLenBoundsNotFound

AllAffiliateCountersCollectionLenBoundsHasGenerics

AllAffiliateCountersCollectionLenBoundsWrongIdent

AllAffiliateCountersCollectionLenBoundsBoundNotFound

CounterIdentCollectionSliceNotFound

CounterIdentCollectionSliceWrongIdent

CounterIdentCollectionSliceWrongType

CounterIdentCollectionSliceHasGenerics

CounterIdentCollectionSliceHasExpr

CounterIdentCollectionGenArrayNotFound

CounterIdentCollectionGenArrayWrongIdent

CounterIdentCollectionGenArrayWrongType

CounterIdentCollectionGenArrayHasGenerics

CounterIdentCollectionGenArrayHasExpr

CounterIdentHashCollectionSliceNotFound

CounterIdentHashCollectionSliceWrongIdent

CounterIdentHashCollectionSliceWrongType

CounterIdentHashCollectionSliceHasGenerics

CounterIdentHashCollectionSliceHasExpr

CounterIdentHashCollectionGenArrayNotFound

CounterIdentHashCollectionGenArrayWrongIdent

CounterIdentHashCollectionGenArrayWrongType

CounterIdentHashCollectionGenArrayHasGenerics

CounterIdentHashCollectionGenArrayHasExpr

CounterIdentHistoricalCollectionLenConstNotFound

CounterIdentHistoricalCollectionLenConstWrongIdent

CounterIdentHistoricalCollectionLenConstWrongType

CounterIdentHistoricalCollectionLenConstHasGenerics

CounterIdentHistoricalCollectionLenConstHasExpr

CounterIdentHistoricalCollectionLenTypeNumNotFound

CounterIdentHistoricalCollectionLenTypeNumHasGenerics

CounterIdentHistoricalCollectionLenTypeNumWrongIdent

CounterIdentHistoricalCollectionLenTypeNumBoundNotFound

CounterIdentHistoricalCollectionSliceNotFound

CounterIdentHistoricalCollectionSliceWrongIdent

CounterIdentHistoricalCollectionSliceWrongType

CounterIdentHistoricalCollectionSliceHasGenerics

CounterIdentHistoricalCollectionSliceHasExpr

CounterIdentHistoricalCollectionGenArrayNotFound

CounterIdentHistoricalCollectionGenArrayWrongIdent

CounterIdentHistoricalCollectionGenArrayWrongType

CounterIdentHistoricalCollectionGenArrayHasGenerics

CounterIdentHistoricalCollectionGenArrayHasExpr

CounterIdentHashHistoricalCollectionSliceNotFound

CounterIdentHashHistoricalCollectionSliceWrongIdent

CounterIdentHashHistoricalCollectionSliceWrongType

CounterIdentHashHistoricalCollectionSliceHasGenerics

CounterIdentHashHistoricalCollectionSliceHasExpr

CounterIdentHashHistoricalCollectionGenArrayNotFound

CounterIdentHashHistoricalCollectionGenArrayWrongIdent

CounterIdentHashHistoricalCollectionGenArrayWrongType

CounterIdentHashHistoricalCollectionGenArrayHasGenerics

CounterIdentHashHistoricalCollectionGenArrayHasExpr

CounterIdentExpectedHashCheckerNotFound

CounterIdentExpectedHashCheckerWrongIdent

CounterIdentExpectedHashCheckerNotUnitType

CounterIdentExpectedHashCheckerHasGenerics

CounterIdentExpectedHashCheckerHasExpr

Implementations

impl IdentTraitItemsBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.