instances_macros::traits::errors

Enum MetaTraitItemsBugs

pub(super) enum MetaTraitItemsBugs {

    CountersTyMetaNotFound {},
    CountersTyMetaWrongIdent {},
    CountersTyMetaExprNotFound {},
    CountersTyMetaInvalidExpr {},
    CountersTyMetaInvalidType {},
    CountersTyMetaHasGenerics {},
    CountersTyNotFound {},
    CountersTyWrongIdent {},
    CountersTyHasGenerics {},
    CountersTyHasBounds {},
    CountersTyCheckerWrongIdent {},
    CountersTyCheckerNotUnitType {},
    CountersTyCheckerHasExpr {},
    CountersTyCheckerNotFound {},
    CountersTyCheckerHasGenerics {},
    CountersGenericsMetaNotFound {},
    CountersGenericsMetaWrongIdent {},
    CountersGenericsMetaExprNotFound {},
    CountersGenericsMetaInvalidExpr {},
    CountersGenericsMetaInvalidType {},
    CountersGenericsMetaHasGenerics {},
    CountersGenericsIndexesMetaNotFound {},
    CountersGenericsIndexesMetaWrongIdent {},
    CountersGenericsIndexesMetaHasExpr {},
    CountersGenericsIndexesMetaInvalidType {},
    CountersGenericsIndexesMetaHasGenerics {},
    CountersLenMetaNotFound {},
    CountersLenMetaWrongIdent {},
    CountersLenMetaInvalidType {},
    CountersLenMetaHasGenerics {},
    CountersLenMetaHasExpr {},
    CountersGenericsCheckerWrongIdent {},
    CountersGenericsCheckerNotUnitType {},
    CountersGenericsCheckerHasExpr {},
    CountersGenericsCheckerNotFound {},
    CountersGenericsCheckerHasGenerics {},
    InstanceTraitDocsRawAttrMissing {},
    InstanceTraitDocsMacroInlineAttrMissing {},
    InstanceTraiDocsNotDocsAttr {},
    InstanceTraiDocsAttrNotLiteralExpr {},
    InstanceTraiDocsAttrNotStrLiteral {},
    InstanceTraitDocsInvalidRawAttr {},
    InstanceTraitDocsInvalidMacroInlineAttr {},
    OriginalCounterConstNotFound {},
    OriginalCounterConstWrongIdent {},
    OriginalCounterConstHasExpr {},
    OriginalCounterConstWrongType {},
    OriginalCounterConstHasGenerics {},
    CounterConstAnnotatedExtractionFailed {},
    CumulatedConstCheckerNotFound {},
    CumulatedConstCheckerWrongIdent {},
    CumulatedConstCheckerHasExpr {},
    CumulatedConstCheckerNotUnitType {},
    CumulatedConstCheckerHasGenerics {},
    CountersParamMetaExtractionGenericsMetaNotFound {},
    CountersParamMetaExtractionCountersTyMetaNotFound {},
    CountersParamMetaExtractionCountersTyMetaDefaultExprNotFound {},
    CountersParamMetaExtractionCountersTyMetaDefaultExprNotLitExpr {},
    CountersParamMetaExtractionCountersTyMetaDefaultExprNotStrLit {},
    CountersParamMetaExtractionGenericsMetaDefaultExprNotFound {},
    CountersParamMetaExtractionGenericsMetaDefaultExprNotArray {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemNotTuple {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleWrongLen {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotFound {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotExprLit {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotLitInt {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotFound {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotExprLit {},
    CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotLitStr {},
    CountersParamMetaExtractionInconsistentWithGenericIdentArgs {},
    CountersParamMetaExtractionInconsistentWithGenericIndexArgs {},
    WhereClauseNotFoundForMarkerBound {},
    MarkerBoundPredNotFound {},

}

Variants

CountersTyMetaNotFound

CountersTyMetaWrongIdent

CountersTyMetaExprNotFound

CountersTyMetaInvalidExpr

CountersTyMetaInvalidType

CountersTyMetaHasGenerics

CountersTyNotFound

CountersTyWrongIdent

CountersTyHasGenerics

CountersTyHasBounds

CountersTyCheckerWrongIdent

CountersTyCheckerNotUnitType

CountersTyCheckerHasExpr

CountersTyCheckerNotFound

CountersTyCheckerHasGenerics

CountersGenericsMetaNotFound

CountersGenericsMetaWrongIdent

CountersGenericsMetaExprNotFound

CountersGenericsMetaInvalidExpr

CountersGenericsMetaInvalidType

CountersGenericsMetaHasGenerics

CountersGenericsIndexesMetaNotFound

CountersGenericsIndexesMetaWrongIdent

CountersGenericsIndexesMetaHasExpr

CountersGenericsIndexesMetaInvalidType

CountersGenericsIndexesMetaHasGenerics

CountersLenMetaNotFound

CountersLenMetaWrongIdent

CountersLenMetaInvalidType

CountersLenMetaHasGenerics

CountersLenMetaHasExpr

CountersGenericsCheckerWrongIdent

CountersGenericsCheckerNotUnitType

CountersGenericsCheckerHasExpr

CountersGenericsCheckerNotFound

CountersGenericsCheckerHasGenerics

InstanceTraitDocsRawAttrMissing

InstanceTraitDocsMacroInlineAttrMissing

InstanceTraiDocsNotDocsAttr

InstanceTraiDocsAttrNotLiteralExpr

InstanceTraiDocsAttrNotStrLiteral

InstanceTraitDocsInvalidRawAttr

InstanceTraitDocsInvalidMacroInlineAttr

OriginalCounterConstNotFound

OriginalCounterConstWrongIdent

OriginalCounterConstHasExpr

OriginalCounterConstWrongType

OriginalCounterConstHasGenerics

CounterConstAnnotatedExtractionFailed

CumulatedConstCheckerNotFound

CumulatedConstCheckerWrongIdent

CumulatedConstCheckerHasExpr

CumulatedConstCheckerNotUnitType

CumulatedConstCheckerHasGenerics

CountersParamMetaExtractionGenericsMetaNotFound

CountersParamMetaExtractionCountersTyMetaNotFound

CountersParamMetaExtractionCountersTyMetaDefaultExprNotFound

CountersParamMetaExtractionCountersTyMetaDefaultExprNotLitExpr

CountersParamMetaExtractionCountersTyMetaDefaultExprNotStrLit

CountersParamMetaExtractionGenericsMetaDefaultExprNotFound

CountersParamMetaExtractionGenericsMetaDefaultExprNotArray

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemNotTuple

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleWrongLen

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotFound

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotExprLit

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIndexNotLitInt

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotFound

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotExprLit

CountersParamMetaExtractionGenericsMetaDefaultExprArrayElemTupleCounterGenericIdentNotLitStr

CountersParamMetaExtractionInconsistentWithGenericIdentArgs

CountersParamMetaExtractionInconsistentWithGenericIndexArgs

WhereClauseNotFoundForMarkerBound

MarkerBoundPredNotFound

Implementations

impl MetaTraitItemsBugs

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.