instances_macros::traits::utils

Type Alias CounterType

pub(crate) type CounterType = Type;

Extracts the common unsigned integer type shared by all validated instance counters.

Aliased Type

pub(crate) enum CounterType {

    Array(TypeArray),
    BareFn(TypeBareFn),
    Group(TypeGroup),
    ImplTrait(TypeImplTrait),
    Infer(TypeInfer),
    Macro(TypeMacro),
    Never(TypeNever),
    Paren(TypeParen),
    Path(TypePath),
    Ptr(TypePtr),
    Reference(TypeReference),
    Slice(TypeSlice),
    TraitObject(TypeTraitObject),
    Tuple(TypeTuple),
    Verbatim(TokenStream),

}

Variants

Array(TypeArray)

A fixed size array type: [T; n].

BareFn(TypeBareFn)

A bare function type: fn(usize) -> bool.

Group(TypeGroup)

A type contained within invisible delimiters.

ImplTrait(TypeImplTrait)

An impl Bound1 + Bound2 + Bound3 type where Bound is a trait or a lifetime.

Infer(TypeInfer)

Indication that a type should be inferred by the compiler: _.

Macro(TypeMacro)

A macro in the type position.

Never(TypeNever)

The never type: !.

Paren(TypeParen)

A parenthesized type equivalent to the inner type.

Path(TypePath)

A path like std::slice::Iter, optionally qualified with a self-type as in <Vec<T> as SomeTrait>::Associated.

Ptr(TypePtr)

A raw pointer type: *const T or *mut T.

Reference(TypeReference)

A reference type: &'a T or &'a mut T.

Slice(TypeSlice)

A dynamically sized slice type: [T].

TraitObject(TypeTraitObject)

A trait object type dyn Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

Tuple(TypeTuple)

A tuple type: (A, B, C, String).

Verbatim(TokenStream)

Tokens in type position not interpreted by Syn.